+++ test/all_test.dart
@@ -1,0 +1,11 @@
+import 'plural_test.dart' as plural_test;
+import 'singular_test.dart' as singular_test;
+import 'snake_case_test.dart' as snake_case_test;
+import 'spinal_case_test.dart' as spinal_case_test;
+
+void main() {
+  plural_test.main();
+  singular_test.main();
+  snake_case_test.main();
+  spinal_case_test.main();
+}