+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0
+
+- Handle uncountable nouns.
+- Added singular inflection.
+- Added plural inflection.
+- Refactored to use a dart:convert based API.
+
 ## 0.1.0
 
 - Initial version, convertToSnakeCase / convertToSpinalCase.